[UPDATE]  store 서버 같은 refresh token으로 refresh 요청시 token 비활성화 기능 구현

+++ MessageQueue/store/src/main/java/com/iia/store/service/sign/SignService.java
@@ -90,14 +90,21 @@ public class SignService {
 
     @Transactional
     public UserRefreshTokenResponse refreshToken(String userRefreshToken) {
-        TokenHandler.PrivateClaims userClaims = userRefreshTokenHandler.parse(userRefreshToken).orElseThrow(RefreshTokenFailureException::new);
-        Optional.ofNullable(redisHandler.getValues(userClaims.getId()))
-                .filter(storedToken -> storedToken.equals(userRefreshToken))
+        TokenHandler.PrivateClaims userClaims = userRefreshTokenHandler.parse(userRefreshToken)
+                .orElseThrow(RefreshTokenFailureException::new);
+        String storedToken = Optional.ofNullable(redisHandler.getValues(userClaims.getId()))
                 .orElseThrow(RefreshTokenFailureException::new);
 
+        if (!storedToken.equals(userRefreshToken)) {
+            redisHandler.deleteValues(String.valueOf(userClaims.getId()));
+            throw new RefreshTokenFailureException();
+        }
+
         String newUserAccessToken = userAccessTokenHandler.createToken(userClaims);
-        String newUserRefreshToken = userAccessTokenHandler.createToken(userClaims);
-        redisHandler.deleteValues(String.valueOf(userClaims.getId()));
+        String newUserRefreshToken = userRefreshTokenHandler.createToken(userClaims);
+
+        redisHandler.setValues(String.valueOf(userClaims.getId()), newUserRefreshToken);
+
         return new UserRefreshTokenResponse(newUserAccessToken, newUserRefreshToken);
     }
 }
